+++ src/main/java/roomescape/global/exception/InvalidRequestFormatException.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception.common;
+package roomescape.global.exception;
 
 import roomescape.global.exception.base.BusinessException;
