+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.global.auth.Authorized;
-import roomescape.global.exception.common.InvalidRequestFormatException;
+import roomescape.global.exception.InvalidRequestFormatException;
 import roomescape.reservation.controller.dto.ReservationRequest;
 import roomescape.reservation.controller.dto.ReservationResponse;
 import roomescape.reservation.controller.dto.ReservationUpdateRequest;
