+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.exception.common.InvalidRequestValueException;
+import roomescape.global.exception.InvalidRequestValueException;
 import roomescape.reservation.service.ReservationService;
 import roomescape.theme.controller.dto.ThemeResponse;
 import roomescape.theme.service.ThemeService;
